requirements.txt
Flask==0.10.1
Flask-Assets==0.8
Flask-Cache==0.12
Flask-KVSession==0.6.1
Flask-LazyViews==0.6
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Redis==0.0.6
Flask-SQLAlchemy==0.16
Flask-Script==0.5.3
Flask-Security==1.7.4
Flask-SocketIO==0.4.0
Flask-WTF==0.10.2
Giflike==0.7
Jinja2==2.7.3
M2Crypto==0.21.1
Mako==0.8.1
MarkupSafe==0.23
MySQL-python==1.2.4
PasteDeploy==1.5.0
PyChecker==0.8.18
PyYAML==3.10
Pygments==1.6
Reindent==0.1.1
SQLAlchemy==0.8.1
Scrapy==0.18.4
Sphinx==1.2b1
Twisted==13.2.0
WTForms==2.0.1
WTForms-JSON==0.2.8
WebOb==1.3.1
Werkzeug==0.9.6
alembic==0.5.0
amqp==1.0.12
anyjson==0.3.3
astroid==1.0.1
awscli==1.2.10
bcdoc==0.12.0
billiard==2.7.3.30
blinker==1.3
botocore==0.30.0
bson==0.3.3
celery==3.0.20
closure-linter==2.3.13
colorama==0.2.5
cssmin==0.1.4
ddbmock==1.0.3
distribute==0.7.3
docutils==0.10
enum34==1.0
factory-boy==1.3.0
flask-csrf==0.9.2
flask-mongoengine==0.7.1
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
git-remote-helpers==0.1.0
greenlet==0.4.4
gunicorn==18.0
httpagentparser==1.7.4
ipaddress==1.0.6
isort==3.8.1
itsdangerous==0.24
jmespath==0.2.1
jsmin==2.0.3
kombu==2.5.12
logilab-common==0.60.1
lxml==3.2.4
mock==1.0.1
mongoengine==0.8.7
msgpack-python==0.4.0
natsort==3.2.0
nodeenv==0.11.1
nose==1.3.0
onctuous==0.5.84
passlib==1.6.2
pep8==1.4.6
pies==2.6.1
pies2overrides==2.6.1
ply==3.4
pyOpenSSL==0.13.1
pyasn1==0.1.7
pycrypto==2.6.1
pylint==1.1.0
pymongo==2.7.2
pyramid==1.5a3
python-dateutil==2.1
python-gflags==2.0
pytz==2013.8
pyzmq==14.0.1
queuelib==1.0
redis==2.7.6
repoze.lru==0.6
rsa==3.1.2
salt==0.17.4
simplejson==3.3.0
simplekv==0.9.2
six==1.3.0
sphinxcontrib-httpdomain==1.1.8
translationstring==1.1
vboxapi==1.0
venusian==1.0a8
virtualenv==1.9.1
w3lib==1.3
waitress==0.8.8
webassets==0.8
wsgiref==0.1.2
youtube-dl==2014.05.13
zope.deprecation==4.0.2
zope.interface==4.0.5
